Improve <object> and <img> svg support
DEVSIX-8829

Autoported commit.
Original commit hash: [67c9cfa6f]

+++ itext/itext.html2pdf/itext/html2pdf/css/apply/util/BackgroundApplierUtil.cs
@@ -625,7 +625,7 @@ namespace iText.Html2pdf.Css.Apply.Util {
                     svgRootRenderer.SetAttribute(SvgConstants.Attributes.WIDTH, null);
                     svgRootRenderer.SetAttribute(SvgConstants.Attributes.HEIGHT, null);
                 }
-                svgImageXObject.UpdateBBox(finalWidth, finalHeight);
+                svgImageXObject.UpdateBBox((float)finalWidth, (float)finalHeight);
                 svgImageXObject.Generate(null);
                 return new float[] { (float)finalWidth, (float)finalHeight };
             }
